+++ packages/server/api/routes/index.js
@@ -10,7 +10,7 @@ const categories = require('./categories.router');
 const reviews = require('./reviews.router');
 const favorites = require('./favorites.router');
 const users = require('./users.router');
-const uidreview = require('./uidreview.router');
+
 const swaggerJsDoc = require('swagger-jsdoc');
 const swaggerUi = require('swagger-ui-express');
 
@@ -39,6 +39,6 @@ router.use('/reviews', reviews);
 router.use('/views', views);
 router.use('/crew', crew);
 router.use('/favorites', favorites);
-router.use('/uidreview', uidreview);
+
 module.exports = router;
 router.use('/categories', categories);
